ShadowPen/backend/requirements.txt
fastapi
uvicorn
playwright
httpx
aiohttp
openai
python-dotenv
beautifulsoup4
websockets
